pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cfo-panel"
version = "0.1.0"
description = "Python + dbt starter project for a CFO analytics data mart"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "dbt-bigquery>=1.8,<1.11",
  "pandas>=2.2.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "streamlit>=1.44.0,<2.0.0",
  "uvicorn>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.0.0,<25.0.0",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
